Add more tests and rearrange a few things
